masm: split code to reduce codegen size
