+++ codegen/masm/src/codegen.rs
@@ -1,5 +1,10 @@
 use crate::config::CodegenConfig;
 use crate::constants::{AUX_TRACE, MAIN_TRACE};
+use crate::error::CodegenError;
+use crate::utils::{
+    boundary_group_to_procedure_name, load_quadratic_element, periodic_group_to_memory_offset,
+    quadratic_element_square,
+};
 use crate::visitor::{
     walk_boundary_constraints, walk_integrity_constraint_degrees, walk_integrity_constraints,
     walk_periodic_columns, walk_public_inputs, AirVisitor,
@@ -74,76 +79,6 @@ struct Backend<'ast> {
 
     /// Configuration for the codegen.
     config: CodegenConfig,
-}
-
-/// Given a periodic column group position, returns a memory offset.
-///
-/// Periodic columns are grouped based on their length, this is done so that only a single z value
-/// needs to be cached per group. The grouping is based on unique lengths, sorted from highest to
-/// lowest. Given a periodic group, this function will return a memory offset, which can be used to
-/// load the corresponding z value.
-fn periodic_group_to_memory_offset(group: u32) -> u32 {
-    // Each memory address contains a quadratic field extension element, this makes the code to
-    // store/load the data more efficient, since it is easier to push/pop the high values of a
-    // word. So below we have to multiply the group by 2, to account for the zero padding, and add
-    // 1, to account for the data being at the low and not high part of the word.
-    group * 2 + 1
-}
-
-/// Loads the `element` from a memory range starting at `base_addr`.
-///
-/// This function is used to load a qudratic element from memory, and discard the other value. Even
-/// values are store in higher half of the word, while odd values are stored in the lower half.
-fn load_quadratic_element(
-    writer: &mut Writer,
-    base_addr: u32,
-    element: u32,
-) -> Result<(), CodegenError> {
-    let target_word: u32 = element / 2;
-    let address = base_addr + target_word;
-
-    // Load data from memory
-    writer.padw();
-    writer.mem_loadw(address);
-
-    // Discard the other value
-    match element % 2 {
-        0 => {
-            writer.movdn(3);
-            writer.movdn(3);
-            writer.drop();
-            writer.drop();
-        }
-        1 => {
-            writer.drop();
-            writer.drop();
-        }
-        _ => unreachable!(),
-    }
-
-    Ok(())
-}
-
-/// Assumes a quadratic extension field element is at the top of the stack and square it `n` times.
-fn quadratic_element_square(writer: &mut Writer, n: u32) {
-    for _ in 0..n {
-        writer.dup(1);
-        writer.dup(1);
-        writer.ext2mul();
-    }
-}
-
-fn boundary_group_to_procedure_name(
-    trace: TraceSegmentId,
-    domain: ConstraintDomain,
-) -> &'static str {
-    match (trace, domain) {
-        (MAIN_TRACE, ConstraintDomain::FirstRow) => "compute_boundary_constraints_main_first",
-        (MAIN_TRACE, ConstraintDomain::LastRow) => "compute_boundary_constraints_main_last",
-        (AUX_TRACE, ConstraintDomain::FirstRow) => "compute_boundary_constraints_aux_first",
-        (AUX_TRACE, ConstraintDomain::LastRow) => "compute_boundary_constraints_aux_last",
-        _ => panic!("Invalid boundary constraint"),
-    }
 }
 
 impl<'ast> Backend<'ast> {
@@ -868,22 +803,6 @@ impl<'ast> Backend<'ast> {
         self.writer
             .mem_load(self.config.trace_domain_generator_address);
     }
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum CodegenError {
-    #[error("invalid access type")]
-    InvalidAccessType,
-    #[error("invalid row offset")]
-    InvalidRowOffset,
-    #[error("invalid size")]
-    InvalidSize,
-    #[error("invalid index")]
-    InvalidIndex,
-    #[error("invalid boundary constraint")]
-    InvalidBoundaryConstraint,
-    #[error("invalid integrity constraint")]
-    InvalidIntegrityConstraint,
 }
 
 impl<'ast> AirVisitor<'ast> for Backend<'ast> {
