+++ codegen/masm/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod codegen;
 mod config;
 pub mod constants;
+pub mod error;
 mod utils;
 pub mod visitor;
 mod writer;
